+++ src/main/java/ar/edu/itba/it/paw/repository/RestaurantRepo.java
@@ -16,7 +16,7 @@ public class RestaurantRepo extends AbstractHibernateRepo {
 		super(sessionFactory);
 	}
 	
-	public Restaurant getRestaurant(long id) {
+	public Restaurant getRestaurant(int id) {
 		return get(Restaurant.class, id);
 	}
